feat: Implement email uniqueness check with error handling on user creation
Created a UserRestController to handle the verification of email uniqueness when adding a new user.
Added corresponding methods in the UserRepository and UserService to check if the email entered during user creation is unique in
the system.
The functionality now displays a modal message indicating that the user with the email "...@..." already exists if the email is already taken. If the email is unique, the user is successfully created.
In case of other errors during the process, appropriate modals are displayed based on whether the post request or the server response failed.
This enhancement ensures a smooth and informative user experience when creating new accounts.