feat: Add functionality to toggle user status (enabled/disabled) with unit tests
Implemented the functionality to update the "enabled" status of a user, allowing it to be toggled between enabled and disabled states.
When clicking the corresponding button, the user's status is updated accordingly.
A notification message displaying the user's ID and the new status (enabled/disabled) is shown at the top of the page using RedirectAttributes.
The changes are successfully persisted in the database.