feat: Add form for creating new categories
Designed a form to facilitate the addition of new categories.
Reused the method for printing subcategories to display them in the select dropdown of the form.
Included inputs for Name and Alias, a select dropdown for Parent Category selection, an image upload button for adding photos, and an enabled checkbox for category activation.