feat: Add category filtering functionality
Implemented category filtering feature with an input field and buttons for searching and clearing the filter.
The filtering logic was successfully integrated, and the functionality was modularized by moving it to the fragments.html file to ensure reusability across both users.html and categories.html pages.
Additionally, resolved an issue where null was being passed as pageNum to the URL during pagination.