feat: Add Brand entity and BrandRepository
Created the Brand entity in the com.shopme.common.entity package along with its corresponding repository, BrandRepository, in the com.shopme.admin.brand package.
The application was executed to generate the respective tables in the database (brands and brands_categories).

The Brand entity establishes a unidirectional many-to-many relationship with the Category entity, with Brand referencing Category.